+++ order-service/src/main/java/com/rushcrew/order_service/infrastructure/messaging/event/OrderCreatedEvent.java
@@ -9,7 +9,7 @@ public record OrderCreatedEvent(
 	Long userId,
 	String status,
 	BigDecimal totalAmount,
-	BigDecimal pointUsed,
+	Long pointUsed,
 	BigDecimal finalAmount,
 	Instant orderedAt
 ) {}
